aula_03_0/requirements.txt
numpy~=2.3.2
glfw~=2.9.0
PyOpenGL~=3.1.9